利息使用双精度(double)

+++ CrazeMM/mine/mortgage/MortgageDTO.h
@@ -10,7 +10,8 @@
 //{"interestRate":1.0E-4,"depotName":"良晋总仓库","quantity":1,"gid":1732,"gimage":null,"updateTime":"2016-06-29 23:13:14","duration":30,"outPrice":101,"checkUid":null,"checkTime":null,"goodName":"锤子-AAAA DV 43 FDD","createTime":"2016-06-29 23:13:14","debtMoney":null,"price":2000.00,"stockId":220,"stateLabel":"待入库","id":23,"state":100}
 @interface MortgageDTO : BaseListDTO
 
-@property (nonatomic) float interestRate;
+@property (nonatomic) double interestRate;
+
 @property (nonatomic) NSInteger quantity;
 @property (nonatomic) NSInteger gid;
 @property (nonatomic) NSInteger duration;
